# perf(lint): cache the returned regex result