### What

The lint rule `no-html-link-for-pages` took a lot of time, as `getUrlFromPagesDirectories` is pretty heavy on computing that we re-regenerate the regex group each time when we lint the file.


Closes [NDX-662](https://linear.app/vercel/issue/NDX-662/investigate-eslint-performance-impact-of-nextno-html-link-for-pages)


Significantly improved with a giant repo, tested locally on my Mac:

After:
```
@next/next/no-html-link-for-pages                          |   154.328 |     0.4% # 1st result
@next/next/no-html-link-for-pages                          |   130.555 |     0.4% # 2nd result
@next/next/no-html-link-for-pages                          |    89.901 |     0.3% # 3rd result
```

Before:

```
(4296 of 4859 files processed...)
Rule                                                       | Time (ms) | Relative
:----------------------------------------------------------|----------:|--------:
@next/next/no-html-link-for-pages                          | 10833.462 |    26.1% # 1st result
@next/next/no-html-link-for-pages                          | 11371.111 |    23.1% # 2nd result
@next/next/no-html-link-for-pages                          | 10944.887 |    21.3% # 3rd result
```